backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
email-validator
redis
sentry-sdk[fastapi]
